AMD doesn't like kernel arguments with types bool/half/size_t/ptrdiff_t/intptr_t/uintptr_t/pointer-to-pointer (!)

+++ src/main/scala/scalacl/impl/data/ScheduledBuffer.scala
@@ -49,7 +49,7 @@ object ScheduledBuffer {
   private val clearBytesKernel = new KernelDef(
     """
     kernel void f(global char* buffer, long length) {
-      size_t i = get_global_id(0);
+      long i = get_global_id(0);
       if (i >= length) return;
       buffer[i] = 0;
     }
